+++ buildSrc/src/main/kotlin/AndroidTestExtRunner.kt
@@ -1,0 +1,6 @@
+object AndroidTestExtRunner : Dependency {
+
+    override val group = "androidx.test.ext"
+    override val artifact = "junit"
+    override val version = "1.1.1"
+}