+++ src/commonMain/kotlin/com/algolia/search/endpoint/EndpointMultiCluster.kt
@@ -99,4 +99,28 @@ public interface EndpointMultiCluster {
         query: UserIDQuery = UserIDQuery(),
         requestOptions: RequestOptions? = null
     ): ResponseSearchUserID
+
+
+    /**
+     * Assign or move [userIDs] to a [clusterName].
+     * The time it takes to migrate (move) a user is proportional to the amount of data linked to each [UserID].
+     *
+     * @param userIDs List of [UserID] to save
+     * @param clusterName The [ClusterName]
+     * @param requestOptions Configure request locally with [RequestOptions].
+     */
+    suspend fun assignUserIds(
+        userIDs: List<UserID>,
+        clusterName: ClusterName,
+        requestOptions: RequestOptions? = null
+    ): Creation
+
+    /**
+     * @param retrieveMapping If set to true, retrieves [ResponseHasPendingMapping.clusters].
+     * @param requestOptions Configure request locally with [RequestOptions].
+     */
+    suspend fun hasPendingMapping(
+        retrieveMapping: Boolean = false,
+        requestOptions: RequestOptions? = null
+    ): ResponseHasPendingMapping
 }