+++ src/commonMain/kotlin/com/algolia/search/model/rule/Condition.kt
@@ -10,8 +10,8 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 public data class Condition(
-    @SerialName(KeyAnchoring) val anchoring: Anchoring,
-    @SerialName(KeyPattern) val pattern: Pattern,
+    @SerialName(KeyAnchoring) val anchoring: Anchoring? = null,
+    @SerialName(KeyPattern) val pattern: Pattern? = null,
     /**
      * Rule context (format: [A-Za-z0-9_-]+). When specified, the rule is contextual and applies only when the
      * same context is specified at query time (using the ruleContexts parameter). When absent, the rule is generic